Refactor index system to avoid replicating common methods
