+++ package.json
@@ -7,8 +7,8 @@
   "dependencies": {
     "async": "^1.5.0",
     "async-lock": "^0.3.8",
-    "aureooms-js-itertools": "^1.3.0",
     "babel-polyfill": "^6.16.0",
+    "aureooms-js-itertools": "^1.3.0",
     "base64url": "1.0.x",
     "blueimp-md5": "^2.3.1",
     "body-parser": "^1.18.3",
